requirements.txt
# A-Qlegal 2.0 - Core Dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.6.0

# NLP and text processing
spacy>=3.6.0
nltk>=3.8.1
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Search and retrieval
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
chromadb>=0.4.15

# API and web framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Data processing
jsonlines>=3.1.0
tqdm>=4.65.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Visualization and UI
streamlit>=1.25.0
plotly>=5.15.0
reportlab>=4.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pdfplumber>=0.10.0

# Multilingual Support
deep-translator>=1.11.0
langdetect>=1.0.9

# Evaluation Metrics
rouge-score>=0.1.2
sacrebleu>=2.3.0
bert-score>=0.3.13
textstat>=0.7.3

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
typer>=0.9.0
rich>=13.7.0

# Data Download
huggingface-hub>=0.19.0
gdown>=4.7.1

# Additional Utilities
pillow>=10.1.0
matplotlib>=3.8.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
easyocr>=1.7.0
pytesseract>=0.3.10
scikit-image>=0.21.0
face-recognition>=1.3.0